requirements.txt
absl-py==0.13.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2024.12.14
charset-normalizer==2.0.4
choreographer==1.0.9
chumpy==0.70
click==8.1.7
# clip==1.0
colorama==0.4.6
colorlog==6.8.0
comm==0.2.2
cycler==0.10.0
dataclasses-json==0.6.3
datasets==3.1.0
debugpy==1.8.12
decorator==5.0.9
dill==0.3.8
distinctipy==1.3.4
distro==1.9.0
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
fastapi==0.109.2
fastjsonschema==2.21.1
ffmpy==0.3.1
filelock==3.13.1
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.1
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==1.35.0
google-auth-oauthlib==0.4.5
GPUtil==1.4.0
gradio==4.17.0
gradio-client==0.9.0
gradio-multimodalchatbot==0.0.1
greenlet==3.0.3
grpcio==1.48.2 # grpcio==1.39.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.30.2
# human-body-prior==2.2.2.0(일단 완)
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.9
importlib-metadata==8.5.0
importlib-resources==6.1.1
ipdb==0.13.9
ipykernel==6.29.5
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.1.2
joblib==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.6.3
jupyter-core==5.7.2
kaleido==1.0.0
kiwisolver==1.3.1
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.1.10
langsmith==0.0.80
lightning-utilities==0.10.1
lmdb==1.6.2
logistro==1.1.0
mapbox-earcut==1.0.1
Markdown==3.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.4.3
matplotlib-inline==0.1.2
mdurl==0.1.2
moviepy==0.2.3.1
mpl-axes-aligner==1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.37.1
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numpy==1.21.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.1
omegaconf==2.3.0
openai==0.28.0
opencv-python==4.9.0.80
orjson==3.10.15
packaging==23.2
pandas==1.3.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
plotly==6.2.0
prompt-toolkit==3.0.20
protobuf==3.19.6
psutil==5.9.7
ptyprocess==0.7.0
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==2.5.3
pydantic-core==2.14.6
pyDeprecate==0.3.2
pydub==0.25.1
pyglet==2.0.10
pygments==2.17.2
PyOpenGL==3.1.0
pyparsing==2.4.7
pyrender==0.1.45
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.7
pytorch-lightning==2.1.3
pytz==2021.1
PyYAML==5.4.1
pyzmq==26.2.0
referencing==0.33.0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.7.2
ruff==0.2.1
safetensors==0.4.1
scikit-learn==0.24.2 # 만약에 scikit-image==0.19.3 에러나면 보셈
scipy==1.7.1
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==1.39.1
setproctitle==1.3.3
shapely==2.0.2
shellingham==1.5.4
shortuuid==1.0.11
simplejson==3.20.1
six==1.17.0
sklearn==0.0
smmap==5.0.1
smplx==0.1.28
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.36.3
sympy==1.12
tbparse==0.0.9
tenacity==8.2.3
tensorboard==2.6.0 
# tensorboard==2.12.0
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
timm==0.3.4
tokenizers==0.15.0
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
# torch==2.4.1+cu121(O)
torchmetrics==0.7.0
# torchvision==0.19.1+cu121(O)
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.36.2
trimesh==4.0.9
triton==3.0.0
typer==0.9.0
typing-extensions==4.9.0
typing-inspect==0.9.0
urllib3==1.26.18
uvicorn==0.27.0.post1
wandb==0.16.1
wcwidth==0.2.5
websockets==11.0.3
Werkzeug==2.0.1
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.2
